use existing raw image if present
somehow the old approach of reading, but not using, the DynamicImage
(when all shrunken images were already present and the raw image was
just copied) was causing a memory leak.

this approach still 'produces' a DynamicImage as part of the
readImageWithMetadata return value.  but since we never use it, lazy
evaluation now seems able to figure out either that it can never be
loaded, or that it can be immediately GC'd.

I don't claim to completely understand why this works, but it does.