fix back-button bugs by removing tracking of scroll position in url
this turns out to be unnecessary once all InternalUrlClicked ->
SetAlbumPathFromUrl navigation has the necessary conversion to "local"
/ "targeted" messages implemented, so that scroll position tracking
can happen purely via the scrolling-related fields in LoadedList &
LoadedAlbum.

and it turns out to cause spurious url changes as scroll positions are
loaded, restored, zeroed out, etc., which can break the back and
forward buttons.

this does mean that we no longer preserve the scroll position on page
reload, but that seems like a minor feature that nobody will miss, and
certainly not worth the trouble of all the hard-to-debug spurious url
change bugs.

there may be more code along these lines that can be removed, but
these are the basics necessary to eliminate the above-mentioned bugs,
plus removal of all the debouncing that was needed to avoid the Safari
issue mentioned in the removed comment.