client: search upwards for album data, using discarded path segments for navigation
in a preceding commit, gen-album began placing symlinks to the
front-end app at every depth in the album.  this means that we may now
be loaded from a url that does not have 'album.json' at a sibling url,
but up at some (great-)*uncle url.

this means when we get a 404 on album.json, we must keep looking if
we're not at the root path yet.  so, we add handling in the 'NoAlbum'
case for this, which chops off one segment of the path, stores it for
later navigation use, and fires off a new getAlbumDataCmd.

this requires the introduction of a new AlbumPath type that captures
whether the current albumPathsAfterLoad came from an old-style
fragment or a new-style path segment.

we also update mavToMsg to attempt to parse an AlbumPath out of both
an old-style fragment and a new-style path segment (relative to the
current baseUrl).  (we prefer a hash path if both are present.)  and
update parsePath to filter out empty path segments, to avoid confusing
navToMsg.