client/Main.locFor: set path rather than fragment
with the handling for path segments in the parent commit, we can now
create links that use them rather than putting the path in the
fragment.

almost all of this is just replacing "hash" with "path" in various
function, parameter, and variable names.  the substantive change is to
use 'NewPath' rather than 'NewFragment' at the very end of the locFor
function.