client: ensure comparisons of paths to titles agree on percent-encoded-ness
attempts to dereference the (percent-encoded) path in the URL against
the (non-percent-encoded) album and image titles need to make sure to
compare only non-percent-encoded strings.  introducing an explicit
PercentEncoded type helps ensure this is the case.

the key missing piece was that, in the 404 sub-case of
Main.updateBootstrap's NoAlbum case, we weren't percent-decoding the
lastPathSegment before storing it in albumPathsAfterLoad.